restructure key and state provisioning server
