+++ enclave-runtime/src/lib.rs
@@ -100,7 +100,8 @@ pub mod cert;
 pub mod error;
 pub mod rpc;
 mod sync;
-pub mod tls_ra;
+pub mod tls_ra_client;
+pub mod tls_ra_server;
 pub mod top_pool_execution;
 
 #[cfg(feature = "test")]
